+++ addon/objects/uml-primitives/fd-uml-class.js
@@ -317,11 +317,19 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }.bind(this));
 
     this.$box.find('.class-stereotype-input').on('blur', function(evt) {
-      this.showNormalizedStereotypeOnInput($(evt.target));
+      let stereotypeText = $(evt.target).val();
+      let stereotype = this.normalizeStereotype(stereotypeText);
+      let rows = stereotypeText.split(/[\n\r|\r|\n]/);
+      let $stereotypeInput = this.$box.find('.class-stereotype-input');
+      $stereotypeInput.val(stereotype);
+      $stereotypeInput.prop('rows', rows.length);
+      let objectModel = this.model.get('objectModel');
+      objectModel.set('stereotype', stereotype);
+      this.paper.trigger('updaterepobj', objectModel, 'stereotype', stereotype);
+      this.updateRectangles();
     }.bind(this));
 
     this.updateInputValue();
-    this.showNormalizedStereotypeOnInput(this.$box.find('.class-stereotype-input'));
 
     // Update the box position whenever the underlying model changes.
     this.model.on('change', this.updateBox, this);
@@ -517,14 +525,5 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }
 
     return stereotype;
-  },
-
-  showNormalizedStereotypeOnInput(element) {
-    let stereotypeText = element.val();
-    let stereotype = this.normalizeStereotype(stereotypeText);
-    let rows = stereotypeText.split(/[\n\r|\r|\n]/);
-    element.val(stereotype);
-    element.prop('rows', rows.length);
-    this.updateRectangles();
   }
 });