+++ addon/objects/uml-primitives/fd-uml-partition.js
@@ -4,10 +4,10 @@
 
 import { computed } from '@ember/object';
 import { isArray } from '@ember/array';
+import joint from 'npm:jointjs';
 
 import FdUmlElement from './fd-uml-element';
 import { BaseObject } from './fd-uml-baseobject';
-import joint from 'npm:jointjs';
 
 /**
   An object that describes a partition on the UML diagram.
@@ -83,6 +83,7 @@ export let Partition = BaseObject.define('flexberry.uml.Partition', {
       this.toBack({ deep: true });
     });
   },
+
   getRectangles() {
     return [
       { type: 'header', element: this }