+++ addon/components/fd-uml-diagram-editor.js
@@ -2,7 +2,7 @@ import Component from '@ember/component';
 import { inject as service } from '@ember/service';
 import { A, isArray } from '@ember/array';
 import { computed, get } from '@ember/object';
-import { isNone, isBlank } from '@ember/utils';
+import { isNone, isBlank, isEmpty } from '@ember/utils';
 import { next } from '@ember/runloop';
 import $ from 'jquery';
 import fade from 'ember-animated/transitions/fade';
@@ -59,6 +59,14 @@ FdPopupActions, {
     @type Array
   */
   interactionElements: A(),
+
+  /**
+    Some primitives, like a swimline, can place only in other primitives. This array contains types of available parent elements
+
+    @property parentElements
+    @type Array
+  */
+  parentElements: A(),
 
   /**
    Array items with empty reference count.
@@ -129,6 +137,20 @@ FdPopupActions, {
     }
 
     return true;
+  }),
+
+  /**
+    Flag indicates when on primitive panel choosen child type object.
+
+    @property isCreatedObjectChild
+    @type Bool
+  */
+  isCreatedObjectChild: computed('type', 'parentElements', function() {
+    let isObject = (this.get('type') === 'Object');
+    let parentElements = this.get('parentElements');
+    let isHasParents = !isEmpty(parentElements) && !isNone(parentElements) && isArray(parentElements);
+
+    return isObject && isHasParents;
   }),
 
   /**
@@ -244,16 +266,27 @@ FdPopupActions, {
       if (!isNone(type)) {
         let x = options.x;
         let y = options.y;
-        if (type === 'Object') {
-          let jointjsCallback = this.get('jointjsCallback');
-          let newObject = jointjsCallback(x, y);
-          this.clearData();
 
-          return newObject;
-        } else {
-          let newLink = this.get('newLink');
-          if (type === 'Link' && newLink && !isNone(newLink.getSourceElement())) {
-            newLink.insertVertex(-1, { x: x, y: y });
+        switch (type) {
+          case 'Object': {
+            // Check if now select not child object. If is not, then create object primitive on blank.
+            if (!this.get('isCreatedObjectChild')) {
+              let jointjsCallback = this.get('jointjsCallback');
+              let newObject = jointjsCallback(x, y);
+              this.clearData();
+    
+              return newObject;
+            }
+
+            break;
+          }
+          case 'Link': {
+            let newLink = this.get('newLink');
+            if (newLink && !isNone(newLink.getSourceElement())) {
+              newLink.insertVertex(-1, { x: x, y: y });
+            }
+
+            break;
           }
         }
       }
@@ -420,6 +453,29 @@ FdPopupActions, {
      */
     collapseEditPanelToolbar() {
       this.toggleProperty('_collapseEditPanelToolbar');
+    },
+
+    /**
+    Handler event createChildObject.
+
+    @method createChildObject
+    @param {Object} options selected joint js element.
+    */
+    createChildObject(options) {
+      let element = options.element;
+      let model = element.model.attributes;
+      let type = model.type;
+      let parentElements = this.get('parentElements');
+
+      if (parentElements.includes(type)) {
+        let x = options.x;
+        let y = options.y;
+        let parentElement = options.element.model.id;
+        let jointjsCallback = this.get('jointjsCallback');
+        let newObject = jointjsCallback(x, y, parentElement);
+        this.clearData();
+        return newObject;
+      }
     }
   },
 
@@ -429,12 +485,14 @@ FdPopupActions, {
     @method createObjectData
     @param {function} jointjsCallback function of creating a new object.
     @param {jQuery.Event} e event.
+    @param {Array} parentElements array of possible parent objects if created object is child object.
   */
-  createObjectData(jointjsCallback, e) {
+  createObjectData(jointjsCallback, e, parentElements) {
     if (!this.get('isLinkAdding')) {
       this._clearProperties();
       this.set('jointjsCallback', jointjsCallback);
       this.set('type', 'Object');
+      this.set('parentElements', parentElements);
       this._changeCurrentTargetElement(e);
     }
   },
@@ -482,6 +540,7 @@ FdPopupActions, {
     this.set('storeCallback', undefined);
     this.set('type', undefined);
     this.set('interactionElements', A());
+    this.set('parentElements', A());
     this.set('newLink', undefined);
   },