+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.14.0-beta01]
+### Fixed
+* Fix setButtonStyles
+
 ## [0.13.0]
 ### Added
 * Add resize and colors to statechart diagram.